+++ src/test/testData/reference/js/KnockoutTemplateReferenceRegistrar/displayAreaMustHaveReference/component.js
@@ -1,0 +1,9 @@
+define(['uiComponent'], function (Component) {
+    'use strict';
+
+    return Component.extend({
+        defaults: {
+            displayArea: 'child<caret>Region'
+        }
+    });
+});